fix: ACTUALLY reduce specificity of Bootstrap exts
So that .mr-7.mr-xl-0 is effectual like .mr-5.mr-xl-0 is.

(The .mr-7 is our extension of Bootstrap. The .mr-5 is pure Bootstrap.)

Effect should be .mr-7 is overwritten by .mr-xl-0 on wide screen.

Effect was .mr-7 was not overwritten by .mr-xl-0 on wide screen.

(An .mr-5 is overwritten by .mr-xl-0 on wide screen.)

Problem seems to be .mr-xl-0 defined before .mr-7.

(Notice .mr-xl-0 is defined after .mr-5.)